cast CURLOPT_VERBOSE flag to long
According the documentation (https://curl.haxx.se/libcurl/c/CURLOPT_VERBOSE.html)
curl_easy_setopt expects that third argument will be long
But our implementation pass it as boolean that led to
some warnings.

To fix it cast to long explicitly

Closes #16

+++ smtp/smtpc.c
@@ -131,7 +131,7 @@ smtpc_set_body(struct smtpc_request *req, const char *body, size_t size)
 void
 smtpc_set_verbose(struct smtpc_request *req, bool curl_verbose)
 {
-	curl_easy_setopt(req->easy, CURLOPT_VERBOSE, curl_verbose);
+	curl_easy_setopt(req->easy, CURLOPT_VERBOSE, (long)curl_verbose);
 }
 
 void
